shrinkg the memory array so that it compiles
usize::MAX is too large to compile and cannot even be accessed! What we really need is the full space representable by u16

+++ src/vm.rs
@@ -1,12 +1,12 @@
 pub struct VM {
-    memory: [u16; usize::MAX],
+    memory: [u16; u16::MAX as usize],
     registers: [u16; 10],
 }
 
 impl VM {
     pub fn new() -> VM {
         VM {
-            memory: [0; usize::MAX],
+            memory: [0; u16::MAX as usize],
             registers: [0; 10],
         }
     }
